% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcv_0327.R
\name{estimate.proportion}
\alias{estimate.proportion}
\title{SSMD: A semi-supervised approach for a robust identification of cell types and deconvolution of mouse transcriptomics data}
\usage{
estimate.proportion <- function(data, lambda = 0.75)
}
\arguments{
\item{data}{input gene expression matrix. MGI gene symbol should be as their row names}

\item{lambda}{threshold of mean correlation to define rank-1 co-expression module}
}
\value{
An object of class is also invisibly returned. This is a list containing the following components:
\item{Stat_all}{statistics for all rank-1 co-expression module. \code{CT}: cell type; mean: mean correlation
inside the module; \code{Core_overlap_number}: Overlap number with core marker list;
\code{Core_overlap_rate}: overlap rate with core marker list; \code{BCV_rank}: bcv rank of the first
base}
\item{module_keep}{modules with the high overlap number with core marker list for each cell type}
\item{proportion}{estimated proportion for each cell type}
}
\description{
This package implements methods for estimate cell type proportion in bulk tissue gene expression data.
}
\examples{
 # Simulate gene expression data for 1000 probes and 6 microarrays.
 # Samples are in two groups
 # First 50 probes are differentially expressed in second group
 # candidate marker list and core markers we found in traing step
 load('IM_markers_20190302_mouse.RData')
 load('Mouse_selected_core_markers.RData')
 # load your own gene expression data
 load('example_bulk.RData')
 aaa=estimate.proportion(data,lambda = 0.8)
}
